modify: Move to the reducer domain subfolder
