+++ src/domain/index.ts
@@ -1,2 +1,3 @@
 export * from './types'
+export * from './reducer'
 export { default as getCardFormSubElement } from './getCardFormSubElement'
