+++ src/domain/reducer/cardListReducer.ts
