+++ src/providers/card/CardProvider.tsx
@@ -2,7 +2,7 @@ import { useReducer, PropsWithChildren } from 'react'
 
 import { INITIAL_CARD_STATE } from '@/contants'
 import { CardStateContext, CardDispatchContext } from '@/contexts/card'
-import { cardReducer } from '@/utils'
+import { cardReducer } from '@/domain'
 
 function CardProvider({ children }: PropsWithChildren) {
   const [state, dispatch] = useReducer(cardReducer, INITIAL_CARD_STATE)
