+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export { default as isNumber } from './isNumber'
 export { default as getConvertedStringsByStars } from './getConvertedStringsByStars'
-export * from './reducer'
